chore: update tests to use built in short flag to distinguish between unit tests and e2e, update readme to explain running the tests
