+++ .github/workflows/e2e.yml
@@ -30,4 +30,4 @@ jobs:
           EV_FUNCTION_NAME: ${{ secrets.EV_FUNCTION_NAME_V1 }}
           EV_INITIALIZATION_ERROR_FUNCTION_NAME: ${{ secrets.EV_INITIALIZATION_ERROR_FUNCTION_NAME }}
           EV_SYNTHETIC_ENDPOINT_URL: ${{ secrets.EV_SYNTHETIC_ENDPOINT_URL }}
-        run: go test -v -count=1 -race --tags=e2e ./...
+        run: go test -v -count=1 -race ./...
