+++ e2e/encrypt_e2e_test.go
@@ -1,13 +1,8 @@
-//go:build e2e
-// +build e2e
-
 package e2e_test
 
 import (
-	"os"
 	"testing"
 
-	"github.com/evervault/evervault-go"
 	"github.com/stretchr/testify/require"
 )
 
@@ -326,25 +321,4 @@ func TestE2EEncryptBytesWithDeniedRole(t *testing.T) {
 
 	_, err = client.DecryptByteArray(encrypted)
 	require.Error(t, err)
-}
-
-type MyStruct struct {
-	String string  `json:"string"`
-	Int    int     `json:"int"`
-	Float  float64 `json:"float"`
-	True   bool    `json:"true"`
-	False  bool    `json:"false"`
-}
-
-func GetClient(t *testing.T) *evervault.Client {
-	t.Helper()
-
-	appUUID := os.Getenv("EV_APP_UUID")
-
-	apiKey := os.Getenv("EV_API_KEY")
-
-	client, err := evervault.MakeClient(appUUID, apiKey)
-	require.NoError(t, err)
-
-	return client
 }
