+++ e2e/outbound_e2e_test.go
@@ -1,6 +1,3 @@
-//go:build e2e
-// +build e2e
-
 package e2e_test
 
 import (
